MOTECH-2861: Fixed commcare and openmrs builds after tasks refactor
